Split code into individual files. Cleanup required
