+++ client/client-common.js
@@ -1,0 +1,6 @@
+import { Meteor } from "meteor/meteor";
+
+const SETTINGS = Meteor.settings && Meteor.settings.public &&
+                 Meteor.settings.public.analyticsSettings || {};
+
+export { SETTINGS };
