+++ client/meteor-analytics.js
@@ -1,38 +1,6 @@
 import { Meteor } from "meteor/meteor";
-import { Tracker } from "meteor/tracker";
-import AnalyticsUsers from "./collections.js";
-import analytics from "../vendor/analytics.min.js";
-// TODO: Validate how to import from npm depends
-// import _ from "underscore";
-
-// TODO Refactor to export this as a handy helper when Meteor 1.3 imports/exports are used.
-let userEmail;
-let initialized = false;
-const SETTINGS = Meteor.settings && Meteor.settings.public &&
-                 Meteor.settings.public.analyticsSettings || {};
-
-/*
-* analytics.js might not have loaded it's integrations by the time we start
-* tracking events, page views and identifies.
-* So we can use these *WhenReady() functions to cause the action to be
-* deferred until all the intgrations are ready.
-* TODO consider whether we should export something like this, maybe provide
-* our own api instead of just using analytics.js' api.
-*/
-const trackEventWhenReady = (...args) => {
-  const _args = args;
-  analytics.ready(() => analytics.track.apply(this, _args));
-};
-
-const trackPageWhenReady = (...args) => {
-  const _args = args;
-  analytics.ready(() => analytics.page.apply(this, _args));
-};
-
-const identifyWhenReady = (...args) => {
-  const _args = args;
-  analytics.ready(() => analytics.identify.apply(this, _args));
-};
+import { AnalyticsUsers } from "./collections.js";
+import { trackEventWhenReady, identifyWhenReady } from "./analytics-helpers.js";
 
 /*
 * getUserEmail()
@@ -69,6 +37,12 @@ const getUserEmail = function getUserEmail() {
   return null;
 };
 
+/*
+* trackLogins()
+* track and report Meteor's login & logout events to analytics.
+* TODO: Add support for sign up event. Refactor to remove initialized.
+*/
+let initialized = false;
 const trackLogins = function trackLogins() {
   // Don't run on first time. We need to access Meteor.userId() for reactivity.
   Meteor.userId();
@@ -76,8 +50,8 @@ const trackLogins = function trackLogins() {
     // Ehen Meteor.userId() changes this will run.
     if (Meteor.userId()) {
       // TODO I think it's not guaranteed that userEmail has been set because
-      // the 'analyticsusers' publication might not be ready yet.
-      identifyWhenReady(Meteor.userId(), { email: userEmail });
+      // the 'AnalyticsUsers' publication might not be ready yet.
+      identifyWhenReady(Meteor.userId(), { email: getUserEmail() });
       trackEventWhenReady("Signed in");
     } else {
       trackEventWhenReady("Signed out");
@@ -86,73 +60,4 @@ const trackLogins = function trackLogins() {
   initialized = true;
 };
 
-const _IronRouter = (Package["iron:router"] && Package["iron:router"].Router);
-const _FlowRouter = (Package["kadira:flow-router"] && Package["kadira:flow-router"].FlowRouter) ||
-                    (Package["meteorhacks:flow-router"] && Package["meteorhacks:flow-router"].FlowRouter) ||
-                    (Package["kadira:flow-router-ssr"] && Package["kadira:flow-router-ssr"].FlowRouter) ||
-                    (Package["meteorhacks:flow-router-ssr"] && Package["meteorhacks:flow-router-ssr"].FlowRouter);
-
-if (_FlowRouter && SETTINGS.autorun !== false) {
-  // something context & context.context don't exist, see: #93
-  _FlowRouter.triggers.enter([function (context) {
-    const page = {};
-
-    if (context.path) {
-      page.path = context.path;
-    }
-    if (context.context && context.context.title) {
-      page.title = context.context.title;
-    }
-
-    page.url = window.location.origin + page.path;
-
-    if (context.route && context.route.name) {
-      page.name = context.route.name;
-    } else {
-      page.name = page.path;
-    }
-    if (context.context && context.context.querystring) {
-      page.search = `? ${context.context.querystring}`;
-    } else {
-      page.search = "";
-    }
-    if (_FlowRouter.lastRoutePath) {
-      page.referrer = window.location.origin + _FlowRouter.lastRoutePath;
-    } else {
-      page.referrer = document.referrer;
-    }
-    _FlowRouter.lastRoutePath = page.path;
-
-    trackPageWhenReady(page.name, page);
-  }]);
-}
-
-const initIronRouter = function initIronRouter() {
-  if (_IronRouter) {
-    _IronRouter.onRun(function () {
-      const router = this;
-      Tracker.afterFlush(function () { trackPageWhenReady(router.route.getName()); });
-      this.next();
-    });
-  }
-};
-
-Meteor.startup(() => {
-  if (!_.isEmpty(SETTINGS)) {
-    if (SETTINGS.autorun !== false) {
-      initIronRouter();
-    }
-    analytics.initialize(SETTINGS);
-  } else {
-    console.error("Missing analyticsSettings in Meteor.settings.public");
-  }
-
-  if (Package["accounts-base"]) {
-    Tracker.autorun(function () {
-      userEmail = getUserEmail();
-    });
-    Tracker.autorun(trackLogins);
-  }
-});
-
-export { analytics };
+export { trackLogins, getUserEmail };
